+++ docs/api/changelog.rst
@@ -406,6 +406,8 @@ Version 0.10
   :func:`aioxmpp.security_layer.negotiate_sasl` has been deprecated in favour
   of :class:`aioxmpp.protocol.XMLStream`\ -level handling of timeouts.
 
+* :attr:`aioxmpp.im.conversation.AbstractConversation.service_member`
+
 .. _api-changelog-0.9:
 
 Version 0.9